renamed some class also removed warnings

+++ library/incl/Geode/cocos/base_nodes/CCNode.h
@@ -1140,7 +1140,7 @@ public:
         // start from end for negative index
         if (index < 0) {
             index = -index - 1;
-            for (size_t i = this->getChildrenCount() - 1; i >= 0; i--) {
+            for (unsigned int i = this->getChildrenCount() - 1; i >= 0; i--) {
                 auto obj = geode::cast::typeinfo_cast<T*>(this->getChildren()->objectAtIndex(i));
                 if (obj != nullptr) {
                     if (indexCounter == index) {
@@ -1152,7 +1152,7 @@ public:
             }
         }
         else {
-            for (size_t i = 0; i < this->getChildrenCount(); i++) {
+            for (unsigned int i = 0; i < this->getChildrenCount(); i++) {
                 auto obj = geode::cast::typeinfo_cast<T*>(this->getChildren()->objectAtIndex(i));
                 if (obj != nullptr) {
                     if (indexCounter == index) {
